s_r_s_mrp_subcontracting: compatibility layer
New module to make compatible the reception screen with the
'mrp_subcontracting' module.

With the reception screen we allowed to receive goods move by move (each
time validated), and this new module makes this flow compatible with
subcontracted goods.
